Promod Reflex 1.1b - New Skins and New Guns !
New Skins!
-Replaced the Marines skin with Woodland SAS for better visibility on the standard Promod map pool.

New Guns!
-Skorpion and Barrett added to the Create-a-Class in-game menu in their respective classes!

-Drastically changed MP44 stats to resemble its CoD2 iteration! (Now with one-taps!!!)
-Added P90 as a secret gun, with the stats of the CoD2 Thompson!

After playtesting, fixed many oversights on weapon statistics:

-Nerfed mini-uzi reflex sight LMAOOOO
-Fixed M16 quick switch timing to be in line with normal and suppressed versions
-Removed headbob from all Skorpion variants and muzzleflash from silenced version

Balance Changes:

Skorpion:
-Max damage and min damage ranges both reduced (200 --> 150 and 400 --> 350 respectively)

Barrett:
-Movespeed further reduced from 0.9 --> 0.85
-Quickraisetime increased to 1.25
-Raisetime increased to 1.5
-ADSTransInTime increased from 0.3 --> 0.35

Skorpion felt a little bit too strong up close so hopefully this will reign it in a bit. As well, Barrett was seeming to be far better than the standard scopes, so I tried to change it to feel a bit clunkier so that choosing it is more of a tradeoff between damage and freedom with movement.

Future plans to add a "scout" type of sniper, likely drastically reworking Dragunov or M21 for this. I also would like to rework the G36 in some way that makes it distinct from the M4 Carbine and other weaker ARs. Maybe making it a burst gun? Giving it an ACOG? Maybe both? Need to go back through all reflex guns' stats and make sure they perfectly match their normal counterparts, so that all weapons in the mod are consistent amongst themselves. I'm also still interested in adding red dots to shotguns or acogs to snipers, but I think that is going to be a little more difficult than I thought at first. Could also add final killcam to the mod in the future, that would be poggers.